move dhis2 connection to workspace
