+++ config/schema.py
@@ -10,7 +10,6 @@ from ariadne import (
 )
 from django.contrib.contenttypes.models import ContentType
 
-from hexa.connections.schema import connections_bindables, connections_type_defs
 from hexa.core.schema import config_bindables, config_type_defs
 from hexa.countries.schema import countries_bindables, countries_type_defs
 from hexa.databases.schema import databases_bindables, databases_types_def
@@ -81,7 +80,6 @@ schema = make_executable_schema(
         type_defs,
         identity_type_defs,
         tags_type_defs,
-        connections_type_defs,
         dags_type_defs,
         *s3_type_defs,
         accessmod_type_defs,
@@ -99,7 +97,6 @@ schema = make_executable_schema(
     [
         uuid_scalar,
         opaque_id_scalar,
-        *connections_bindables,
         *pipelines_bindables,
         *pipeline_templates_bindables,
         *identity_bindables,
