--- hexa/connections/graphql/schema.graphql
@@ -1,30 +1,0 @@
-"""
-DHIS2 metadata item
-"""
-type DHIS2MetadataItem {
-    id: String!     # Unique identifier
-    name: String!   # Name of the metadata item
-}
-enum DHIS2ConnectionError {
-    CONNECTION_ERROR
-    UNKNOWN_ERROR
-}
-"""
-DHIS2 metadata query result
-"""
-type DHIS2QueryResult {
-    data: [DHIS2MetadataItem!]  # List of metadata items
-    success: Boolean!           # Query success status
-    errors: [String!]!      # List of errors
-}
-
-"""
-DHIS2 connection object
-"""
-type DHIS2Connection {
-    query(type: String!, filter: String): DHIS2QueryResult! # Query metadata items
-}
-
-extend type Query {
-    dhis2connection(slug: String!): DHIS2Connection # Get DHIS2 connection
-}
