+++ hexa/workspaces/tests/fixtures/indicators.py
